+++ src/decompy/rankmethods/__init__.py
@@ -1,0 +1,14 @@
+from .penalized import rank_elbow, rank_classical_ic, rank_bai_ng, rank_DIC
+from .cvrank import rank_gabriel, rank_separate_rowcol, rank_bcv
+from .bayes import rank_hoffbayes
+
+__all__ = [
+    "rank_elbow",
+    "rank_classical_ic",
+    "rank_bai_ng",
+    "rank_DIC",
+    "rank_gabriel",
+    "rank_separate_rowcol",
+    "rank_bcv",
+    "rank_hoffbayes"
+]