+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - First release of `decompy`
 
+## v0.2.0 (31/12/2023)
+
+- Added matrix rank estimation methods (check `src/decompy/rankmethods`)
+- Added mutiple rank estimation version of rSVDdpd algorithm
+